Account for subclass plugs when matching in-game loadouts.
