+++ src/app/loadout/loadout-edit/LoadoutEditSubclass.tsx
@@ -7,8 +7,8 @@ import { ResolvedLoadoutItem } from 'app/loadout-drawer/loadout-types';
 import { AppIcon, powerActionIcon } from 'app/shell/icons';
 import clsx from 'clsx';
 import { useMemo } from 'react';
+import { getSubclassPlugs } from '../item-utils';
 import EmptySubclass from '../loadout-ui/EmptySubclass';
-import { getSubclassPlugs } from '../loadout-ui/LoadoutSubclassSection';
 import PlugDef from '../loadout-ui/PlugDef';
 import { createGetModRenderKey } from '../mod-utils';
 import styles from './LoadoutEditSubclass.m.scss';
